requirements.txt
PyPDF2==3.0.1
Pillow==10.0.1
tkinter-tooltip==2.1.0
reportlab==4.0.4
pdfplumber==0.9.0
fitz==0.0.1.dev2
PyMuPDF==1.23.8